do not delete external uploads, fixes #424

+++ app/src/main/java/com/soundcloud/android/model/Recording.java
@@ -520,8 +520,8 @@ public class Recording extends ScResource implements Comparable<Recording> {
      */
     public void onUploaded(ContentResolver resolver) {
         upload_status = Status.UPLOADED;
+        if (!external_upload) IOUtils.deleteFile(getFile());
         IOUtils.deleteFile(getEncodedFile());
-        IOUtils.deleteFile(getFile());
         IOUtils.deleteFile(resized_artwork_path);
         updateStatus(resolver);
     }
